+++ src/facades/index.ts
@@ -25,7 +25,15 @@ export {
   isClassicAtwFacade,
   isClassicErvFacade,
 } from './classic-types.ts'
+export {
+  type HomeAtaGroupSource,
+  aggregateClassicAtaGroupStates,
+  toClassicAtaGroupState,
+  toGroupFanSpeed,
+  toHomeAtaValues,
+} from './home-ata-group.ts'
 export { HomeBaseDeviceFacade } from './home-base-device.ts'
+export { HomeBuildingAtaFacade } from './home-building-ata.ts'
 export { HomeDeviceAtaFacade } from './home-device-ata.ts'
 export { HomeDeviceAtwFacade } from './home-device-atw.ts'
 export { HomeFacadeManager } from './home-manager.ts'